Redesign the login signup pages
